Add database handler ref, form submission
The database handle is stored in a reference, because I don't want it
to experience changes or cause re-renders. The component state tracking
database handle loading is set true after the database handle finishes
awaiting the promised IDBPDatabase.

The database handle is then passed to the Weight_workout_Form as a
destructured/named property, and used during form submission.

It works insofar as the happy path is concerned, errors are as yet not
being handled.