Add client-side min-number input validation
I don't really care how a Weight_workout is built or used in the
backend, but as a business-rule observation, it doesn't make sense to be
able to submit a Weight_workout with a value of less than 1. If someone
reaches into the console and submits something like that, I don't have
controls to stop it, but there isn't risk of breaking anything by doing
that, they're just making their own data garbage in the process.