Change function definition to `useCallback()`
React re-runs a component/function each time that a re-render is
invoked, which on the Weight_workout_Form is quite frequently, since the
inputs are controlled values that relate to the Weight_workout object in
memory. This means that every time the Weight_workout_Form function is
invoked, it was doing the work of defining the handle_submit() function,
which is slow. This is what `useCallback()` is built to do: provide a
named memoized function reference that only rebuilds when its
dependencies are updated. In this case, the dependency is on the
database_handle reference, so if that changes out from under us, it
should be handled appropriately. It isn't, yet, but it should be.

This should improve performance.