Modify Weight static factory method for objects
Part 1 of a series of commits to deal with how data gets serialized.
Dead objects have no functions, and ergo, no `get` or `set` accessor or
mutator properties. When data is pulled from an object database, like
IndexedDB, it's dead, Jim. Previous version of the `from_Weight()` factory
method assumed that the public `get` properties may be present, and
returned a new object by filling the constructor with `quantity` and
`weight`, if present. This does me no good if the `get` properties don't
exist but the data squirrelled away in private fields is valid. This now
chains defaulting through the public `get`, then to the private field,
then to the field default.