Change Weight_workout to include UUID
As I thought about the real-world application of the workout tracker, it
occured to me that a UUID would be less likely to have a user-entered
key conflict. I have no immediate plans to implement a change which
would necessitate dealing with this differently, just thinking ahead a
little to achieve good design.