Add Weight_workout factory method from dead object
Again, a decision to deal with database weirdness. Because the objects
are dead when stored in the database, having no function properties, the
standard constructor here will not work as desired, for the same reasons
as Weight's constructor won't: valid properties exist on the object, but
not by the names that are expected or enumerable automatically.

The public properties are the only properties recognized by tyepscript
when using the Partial<Weight_workout> type. This is the desired
behavior on some level, because I want the default constructor to
operate on an object with a number of named fields, and I don't want
to be typing in, or exposing, private field names in order to
instantiate the class.

Rather frustratingly, the `get` and `set` syntax is an all-or-nothing
deal. The `set` function cannot exist without a `get` function which
returns a value of a kind matching its parameter. I need custom behavior
when most of the fields of Weight_workout are mutated, and as such I
need `set`, and therefore, `get` functions. As such, I cannot define a normal
scalar field of the same name which can be read but not written to. I
still have a private backing field, though.

When the object is stored in indexedDB, it dies, and loses its
functions, but keeps its private fields. This is plainly visible in the
devtools. Hence this factory, which being defined as a method of the
same class, has access to the private fields of a dead Weight_workout object.

There are other approaches I could have taken, such as making the fields
I needed `readonly` per typescript parlance, and defining the `set`
functions as `set_fieldname()`, etc, but that's clunky and would force
me to update my code which interacts with this object in a bunch of
places. The factory is the least laborious solution, and keeps
coupling down.