Add fragment <></> wrap, make step smaller, reindent
The fragment element wrapping around things might just be superstition,
as I was trying to solve some compilation bug earlier.

The step on that weight quantity input is because the rounding was
producing weights such as 55.12, and step of 0.1 made that kind of entry
illegal. Scrolling it to change the number isn't very ergonomic with
such a narrow step, but users can type in their number as far as I care.

Then I reindended the whole file. Just one autoformat run from line 1 to
line last