Add Weight_workout_List
Took me... too many hours to get this working. Rather that commit many
small, non-working changes, I opted to commit once it worked.

The async database read needed to have the loading status set inside
its operation, after the final await, so that it would fire a re-render
only *after* everything had been read and written to the list map.

Speaking of map, I ultimately went with a javascript Map type, rather
than a Set or an Array. I tried all of them. An array was impure,
because it would accept as many arbitrary pushes as needed, and react in
development mode runs functions twice. As a result, the array version of
this ended up with lots of duplicates. So did the Set version of this,
it turns out. Javascript Sets are designed so that they only accept
unique values... but objects are considered unique value references,
even if the shallow equality is the same for all fields in two objects.
Set was not as idempotent as I had hoped. But the JS Map type, that
worked, since it doesn't use `push()` or `add()`, but `put()`. I already
had keys I could use, since I was using the id to key the database too,
so that would be just fine.

But only Arrays have the `map()` method, ironically. Both Map and Set
collections have `forEach()`, but that is designed to have no return
value, only side-effects. In order for the pattern of displaying a list
of values to work, the function needs to return JSX elements. That had
me stumped for a while. Ultimately, though, `map()` has a return value,
and `forEach()` does not. But to get `map()` you effectively need to
cast the Map or Set to an Array, which is best accomplished by
constructing a new one ad-hoc.

Okay, then I was getting the data out, and I was logging it, but I was
getting garbage prints, because the database-retrieved objects were
dead. I needed to massage their fields into shape so the newly
instantiated objects got the correct information in its constructor.
Several commits ago is the ground work for that, which happened out of
order in reality, but I'm commiting changes so that dependencies are
appropriately ordered.