Add database handler to achieve score tracking
I basically just copy-pasted a lot of the code from my other
Weight_workout_Database, but it works. Again, taking advantage of idb's
bindings to typescript, I get compile-time checking for my database
according to the datatype I'm trying to wrangle. And the beautiful thing
about interfaces, I don't have to define a whole class implementation to
use the interface as a type definition.

In this case, I am using the Autoincrement feature of indexedDB, which
uses an automatic key and keeps track of that for me. I don't feel a
need to get weird with crypto.randomUUID() like I did for the other
project.