Add database init and writing to App.tsx
Okay, the useEffect() here is hairy as hell. As I allude to in the
comment associated with that call, I was having async race condition
issues. IMO, it didn't make sense from a UX perspective to block
submitting for scores until the database was attached. For one, the
database is a progressive feature, and for another, that'd be a horrific
random wait with no visibility to the user. I also didn't want to
wrangle some event listener the hard way. This is kind of a poor-man's
event listener.

Ultimately, the dependency array in react hooks does not seem to have a
way to AND the dependencies, only ORs, so I do basically have to list
all the dependencies and then to a guard clause in the callback to avoid
extraneous fires.

Seems obvious in hindsight, but it wasn't immediately apparent to me,
because I know that JS does some funky things with the event loop around
async functions. Like, if the event loop can't pop something off the
stack, it's not going to spin up an async worker thread. So a do-until
clause wasn't going to work, because it'd never exit the loop, and never
handle another worker thread returning. I initially tried using an async
function from within the callback, but I couldn't figure out how to
await the truthiness of the database being attached, which I'm sure
there must be *some* way to deal with, but I do not know it. Really what saved
my bacon was just doing console debugging and a little intuition.