Add main and view for learning purposes
Okay, so a little screwing around involved with getting these loaded as
modules, some observations and comments:

In this revision, the state I want to track, the model, is kept local to
the view and directly updated by main. In this experiment, I'm
confirming that when html elements are manipulated in this way, updating
the state directly doesn't cause changes in the DOM. I suspect that what
I have to do is write an update function which either changes the
textContnet or re-renders the whole artifact.