Added update(), fixed loop for count
What we learn here:

For one, setInterval and clearInterval are the correct tools to use for
looping like I was. setTimeout in a loop like that will just set a bunch
of timeouts at once.

Adding the update method to the view was the critical piece here. I'm
not sure about how well this scales, but it'll work.