Add custom event types
Rather than screw around with each of the MVC needing to know about each
other to accomplish an observer pattern, I'm taking advantage of the
event system. Now, I'm not sure if I should be defining events in a
central location as I am here, or if it would be better practice to
define the events near where they are used, such as in the same file as
the module which dispatches them. The problem with this unsurety is that
the purpose is to decouple the components, and that these events
function as an intermediary contract between the components.

Because they effectively function as an agnostic contract, leaving
implementation details up to the modules using them, I suspect that
defining them this way is more forgiving to change.