Rewrite View
Really what I want to be happening here is for the View to have a
handle/pointer/reference to the Model when it is constructed, and to not
need to know anything other than what it is instantiated with, but to
minimize its instantiation requirements. I'm not sure if it matters.
This is sloppy, and an ugly commit, but I basically rewrote the file.

The button I added to the tree in render() references the event defined
in the CUSTOM_EVENTS const I talked about earlier. This allows the View
to notify any interested subscribers without needing to know about them
directly. I think this is ideal, because it lets me define how I want to
handle the button click on the controller without needing to pass a
callback to the View.

I also noodled around with render and update. This is the unclean,
heretical, imperative way of accomplishing this. The update function
really depends on the presence of the stuff set up in the render
function, but rather than have the View call its own render during
instantiation, I thought it better to call it once from the Controller
after other book-keeping, such as subscribing to the
button_clicked_event, was handled.

I might further in this prototype attempt a more clean functional
approach, wherein I build my new element tree and then perform one clean
DOM update to replace the children of the parent element all at once.
This would come at the detriment of requiring more resources for each
state change, but at the benefit of an improved ability to reason about
what happens. Throwing away the old state, to include the rendered out
element tree, and replacing it with the new state is very amenable to
how I like to reason about programs. But it raises questions, such as
for the resource cost: can more precise state changes be accomplished in
a functional manner without trashing the whole element tree?