Rewrite main
Since I learned what I needed to from the interval loop I used earlier,
I'm moving on to a real MVC architecture implementation. The controller,
here Main, is insantiating the View and the Model, and then subscribing
to custom events fired in the View, and handling those events. I'm not
sure if there's a cleaner way to express this idea, or one which is more
idomatic to javascript, but I'm confident there are better ways.