Add tsconfig.json
There are several things of note here

Of greatest import to the environment I'm dealing with here, being
containerized in a virtual machine, is Polling for the watch features.
To that end, the watchOptions section of the config uses the
PriorityPollingInterval, which treats some filetypes as more important
to check more frequently.

The next is the module type. This had one of several choices outlined in
the typescript documentation. Ultimately, because I want to use the
module facilities in the browser, valid for some 95%+ of users right
now, I need to target ES2015 modules. If I were more concerned about a
higher % threshold, I might have chosen to use webpack and commonjs or
some other means of accomplishing this, but I want to reduce
dependencies. I'm aiming to accomplish features, not enterprise-level
user support.