Add json data as javascript module which exports object
Something I was taking for granted in the create-react-app scaffolded
version of this project is that the bundler actually grabbed the JSON
content of the .json file when I was importing it, and just created an
object right there in the file by parsing that literal data. The
browser, however, does not like it if you attempt to load MIME type
application/json data as a javascript module.

This commit just copies the json content into a typescript file and
exports a named constant object. Even better, in this case, because I
actually get the benefits of the typechecking since I'm assigning the
constant object the type of Quiz