Add view
I'm trying to make this as stateless/functional as possible. I'm not
sure how far down that line of thinking I want to go, but even if it's
only on the initial DOM insertions, I want to attempt to deal with
things with as few side-effects as possible. Regardless of the
performance benefits or penalty, I think it's much easier to work with a
DocumentFragment to build a node tree, and then directly append that
node tree all at once. It allows me to scope all the node tree building
to inside functions, which adheres to the idea of avoiding global
mutable state.