Update Main to use a class oriented approach
Ironically I think I was causing myself an unexpected problem by doing
this. idk. I did not know about the issue with binding "this", and I
have a new distaste for javascript. I'd rather that be an opt-in feature
than opt-out, which is very clunky. Got stuck on that for a while.