Add util function to generate CSS safe ids
As noted in the jsdocs around the function, HTML will let you get away
with using ids that CSS does not like. To avoid any weirdness with
element ids not being selectable, this function strips out offending
characters and prefixes the string with "id_", which avoids problems
around ids starting with numbers.