Add curry/closure to support UUID at runtime
Probably I could have split the next several changes out into their own
branch. As I have been going down the path of rendering in a functional
manner, I've been thinking about how to narrow the scope of changes. My
initial thought last night was some convoluted mess of an interface that
takes among its parameters a parent element so that it can use some
variation on `parent.replaceChildren()`. This is wholly unnecessary,
since the DOM api exposes a method on elements to replace that specific
element, `element.replaceWith()`. This has pretty good support according
to caniuse, and it's very slick. In combination with
`element.closest(css_selector)`, I can use element ids in order to grab
an event.target's parent matching an id, and then replace that element
directly with a newly generated fragment built in the same way.

This requires some discipline, in the same way that React jsx syntax
will complain if used without a single canonical parent element to which
the subtree attaches.

Now, for the next few changes I commit here, I could have just hardcoded
an ID in a const at the top of the file and called it a day. But I
wanted to pursue this as a learning exercise, since this is a class
project. The solution here is absolutely overengineered, but it works.

The engineering is such:

A fragment builder, which returns a DocumentFragment, accepts a UUID as
a parameter. A dummy method looks like it and accepts all the same
state, minus the UUID. The curry/closure higher-order-function accepts
the fragment builder as a parameter, and closes around it with a UUID
generated at runtime, returning a new function matching the call
signature of the dummy method, so that every subsequent invocation of
the returned function has the same UUID encoded. The constructor of the
class then reassigns the dummy method to the returned closure function.
The 'true' fragment builder is private, and the dummy method having been
assigned is public. The dummy method can then be called ad infinitum,
and its result, with the identical UUID, used to select the parent
element to `element.replaceWith()`.