Breakout header to section, call form as child
This is primarily an academic exercise, to deal with the same UUID
closure/currying, but with additional children to the fragment builder.
How much work is it really saving the computer by lifting the headers
out to a section? None. It's just not that expensive computationally.

But it *does* follow the discipline required to use the
```
build_fragment() -> new_fragment -> closest(new_fragment.UUID) -> replaceWith(new_fragment)
```
design.

This makes sure that both the form and the section headers have their
own single canonical parent.