Add database handle attachment to Model
The import of idb is resolved completely down to the js file because of
the ES native module handling.

Not seen around the db_handle_ready property is the troubleshooting
performed around the problem of defining a part of the object
asynchronously. I need to `await` the database connection if I want to
assign it in the constructor, and I can't really make the constructor
async... An async IIFE doesn't work to satisfy the typescript compiler
that the property is definitely assigned in the constructor, and I don't
blame it. The constructor needs to return an object of the class, with a
known state.  A promise is in some sense a known state, for a number of
reasons. And, per the SO post, I can use the Promise as effectively
equivalent to a handle, defining what I want to do in a `.then()` call.

This does help avoid some weirdness around async/await functions, since
I can define a `.then()` and keep moving right along without needing to
break a function up into all the crap that happens before an `await` and
all the stuff that depends on that `await`.