Add Model.score_quiz() which writes to the db
Uses the `.then()` property of the db_handle_ready Promise in order to
define side-effects that need to happen as it moves right along. If the
db_handle_ready Promise is already in a ready state, it'll do all that
right then, but if not, it'll happen when it is ready.