Add score_history functionality to the Model
The `get_score_history()` method is async because I needed to return the
score_history_list definitively, and typescript was complaining that the
function had a void return type, so I couldn't use the output of the
function in any meaningful way. Now, maybe I could have returned the
db_handle_promise call with the .then and etc, but I was confused on how
to accomplish this smoothly. Ultimately, this returns a Promise I have
to resolve, but it does return a value.