Add Main method for dealing with score_history_list
Calling it again like that is a little brute-force, but it works.

Note that in the update_history_list() method I'm dealing with the value
of the new list in the Promise.then() handler. This is easy, since there
is no return value and it just invokes a callback with side-effects.

Turns out that whole currying/closure around fragment builders was a
really nifty design that let me just kind of stumble into being able to
do this. Again, granted, I could have defined some constants for element
IDs and futzed with it that way, but this works great.