Make changes to try to address Cumulative Layout Shift
There are conflicting sources and conflicting reports, so this is
ultimately kind of a crapshoot whether or not it works. Lighthouse
randomly does and does not complain of CLS of the body, and of several
constituent elements which have their layout changed by styles. So I
figured I'd put style stuff up front and give the body a min-height so
it wouldn't reflow in a strange way. Even still, lighthouse sometimes
gives me a strange result, where it'll knock me down several points for
CLS, and other times it doesn't find anything to complain about.